all business logic in vm(location)
